use [optional-dependencies] for examples and gsw; [dependency-groups] for tes and docs
- move load_OCCA() from neutralocean.data to examples/run_OCCA.py
- numpy trapz -> trapezoid
